➕ Feat(board/exception): 사용자 권한 예외 처리 (#698)
- CommentServiceImpl
- MeetingServiceImpl
- NoticeServiceImpl
- TalkServiceImpl

+++ backend/src/main/java/com/pi/stepup/domain/board/service/comment/CommentServiceImpl.java
@@ -11,10 +11,14 @@ import com.pi.stepup.domain.board.domain.Talk;
 import com.pi.stepup.domain.board.dto.comment.CommentRequestDto.CommentSaveRequestDto;
 import com.pi.stepup.domain.board.dto.comment.CommentResponseDto.CommentInfoResponseDto;
 import com.pi.stepup.domain.board.exception.BoardNotFoundException;
+import com.pi.stepup.domain.board.exception.CommentNotFoundException;
+import com.pi.stepup.domain.user.constant.UserRole;
 import com.pi.stepup.domain.user.dao.UserRepository;
 import com.pi.stepup.domain.user.domain.User;
 import com.pi.stepup.domain.user.exception.UserNotFoundException;
+import com.pi.stepup.global.error.exception.ForbiddenException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -22,6 +26,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static com.pi.stepup.domain.board.constant.BoardExceptionMessage.BOARD_NOT_FOUND;
+import static com.pi.stepup.domain.board.constant.BoardExceptionMessage.COMMENT_NOT_FOUND;
 import static com.pi.stepup.domain.user.constant.UserExceptionMessage.USER_NOT_FOUND;
 
 @Service
@@ -68,6 +73,8 @@ public class CommentServiceImpl implements CommentService {
     @Transactional
     @Override
     public List<CommentInfoResponseDto> readByBoardId(Long boardId) {
+        boardRepository.findOne(boardId)
+                .orElseThrow(() -> new BoardNotFoundException(BOARD_NOT_FOUND.getMessage()));
         List<Comment> allComments = commentRepository.findByBoardId(boardId);
         return allComments.stream()
                 .map(c -> CommentInfoResponseDto.builder().comment(c).build())
@@ -77,6 +84,20 @@ public class CommentServiceImpl implements CommentService {
     @Transactional
     @Override
     public void delete(Long commentId) {
+        Comment comment = commentRepository.findOne(commentId)
+                .orElseThrow(() -> new CommentNotFoundException(COMMENT_NOT_FOUND.getMessage()));
+
+        // 로그인한 사용자의 정보 가져오기
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//        String loggedInUserId = authentication.getName();
+
+        String loggedInUserId = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+
+        // 로그인한 사용자가 댓글 작성자이거나, 관리자일 경우에만 삭제 허용
+        if (!loggedInUserId.equals(comment.getWriter().getId()) && !UserRole.ROLE_ADMIN.equals(comment.getWriter().getRole())) {
+            throw new ForbiddenException();
+        }
+
         commentRepository.delete(commentId);
     }
 }
